rag-knowledge-base/phase0/backend/requirements.txt
# Phase 0 验证 Sprint 依赖
# 目标：快速跑通 pipeline，不做生产级优化

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12        # 文件上传
pydantic==2.9.2

# LangChain
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.5
langchain-core==0.3.23

# 向量存储
qdrant-client==1.12.1

# PDF 解析
unstructured==0.16.5
pdfplumber==0.11.4

# JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis (Phase 0 可选，先不加)
# redis==5.2.0

# 环境变量
python-dotenv==1.0.1
